feat: merge mocks
