+++ pallets/astar-xcm-benchmarks/src/generic/benchmarking.rs
@@ -144,11 +144,7 @@ mod benchmarks {
         )))
     }
 
-    impl_benchmark_test_suite!(
-        Pallet,
-        crate::generic::mock::new_test_ext(),
-        crate::generic::mock::Test
-    );
+    impl_benchmark_test_suite!(Pallet, crate::mock::new_test_ext(), crate::mock::Test);
 }
 
 pub struct XcmGenericBenchmarks<T>(sp_std::marker::PhantomData<T>);
