+++ pallets/astar-xcm-benchmarks/src/mock.rs
@@ -16,8 +16,67 @@
 // You should have received a copy of the GNU General Public License
 // along with Astar. If not, see <http://www.gnu.org/licenses/>.
 
-use frame_support::{dispatch::Weight, parameter_types, traits::ContainsPair};
+//! A mock runtime for XCM benchmarking.
+
+use crate::{fungible, generic, *};
+use astar_primitives::xcm::ReserveAssetFilter;
+use frame_benchmarking::BenchmarkError;
+use frame_support::{
+    assert_ok, parameter_types,
+    traits::{fungible::ItemOf, AsEnsureOriginWithArg, Everything, Nothing},
+    weights::Weight,
+};
+use frame_system::{EnsureRoot, EnsureSigned};
+
+use core::marker::PhantomData;
+use sp_core::{ConstU32, ConstU64, Get, H256};
+use sp_runtime::{
+    testing::Header,
+    traits::{BlakeTwo256, IdentityLookup},
+    BuildStorage,
+};
 use xcm::latest::prelude::*;
+use xcm_builder::{AllowUnpaidExecutionFrom, FungiblesAdapter, MintLocation, NoChecking};
+
+type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
+type Block = frame_system::mocking::MockBlock<Test>;
+type Balance = u64;
+type AccountId = u64;
+type AssetId = u128;
+
+// For testing the pallet, we construct a mock runtime.
+frame_support::construct_runtime!(
+    pub enum Test where
+        Block = Block,
+        NodeBlock = Block,
+        UncheckedExtrinsic = UncheckedExtrinsic,
+    {
+        System: frame_system::{Pallet, Call, Config, Storage, Event<T>} = 10,
+        Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
+        Assets: pallet_assets::{Pallet, Call, Storage, Event<T>},
+        PolkadotXcmGenericBenchmarks: pallet_xcm_benchmarks::generic::{Pallet},
+        PolkadotXcmFungibleBenchmarks: pallet_xcm_benchmarks::fungible::{Pallet},
+        XcmAssetsBenchmark: fungible::{Pallet},
+        XcmGenericBenchmarks: generic::{Pallet},
+    }
+);
+
+pub struct AccountIdConverter;
+impl xcm_executor::traits::Convert<MultiLocation, u64> for AccountIdConverter {
+    fn convert(ml: MultiLocation) -> Result<u64, MultiLocation> {
+        match ml {
+            MultiLocation {
+                parents: 0,
+                interior: X1(Junction::AccountId32 { id, .. }),
+            } => Ok(<u64 as parity_scale_codec::Decode>::decode(&mut &*id.to_vec()).unwrap()),
+            _ => Err(ml),
+        }
+    }
+
+    fn reverse(acc: u64) -> Result<MultiLocation, u64> {
+        Err(acc)
+    }
+}
 
 // An xcm sender/receiver akin to > /dev/null
 pub struct DevNull;
@@ -52,41 +111,269 @@ impl xcm_executor::traits::OnResponse for DevNull {
     }
 }
 
-pub struct AccountIdConverter;
-impl xcm_executor::traits::Convert<MultiLocation, u64> for AccountIdConverter {
-    fn convert(ml: MultiLocation) -> Result<u64, MultiLocation> {
-        match ml {
-            MultiLocation {
-                parents: 0,
-                interior: X1(Junction::AccountId32 { id, .. }),
-            } => Ok(<u64 as parity_scale_codec::Decode>::decode(&mut &*id.to_vec()).unwrap()),
-            _ => Err(ml),
-        }
-    }
+parameter_types! {
+    pub const BlockHashCount: u64 = 250;
+    pub BlockWeights: frame_system::limits::BlockWeights =
+        frame_system::limits::BlockWeights::simple_max(Weight::from_parts(1024, u64::MAX));
+    pub UniversalLocation: InteriorMultiLocation = Here;
+}
+impl frame_system::Config for Test {
+    type BaseCallFilter = Everything;
+    type BlockWeights = ();
+    type BlockLength = ();
+    type DbWeight = ();
+    type RuntimeOrigin = RuntimeOrigin;
+    type Index = u64;
+    type BlockNumber = u64;
+    type Hash = H256;
+    type RuntimeCall = RuntimeCall;
+    type Hashing = BlakeTwo256;
+    type AccountId = AccountId;
+    type Lookup = IdentityLookup<Self::AccountId>;
+    type Header = Header;
+    type RuntimeEvent = RuntimeEvent;
+    type BlockHashCount = BlockHashCount;
+    type Version = ();
+    type PalletInfo = PalletInfo;
+    type AccountData = pallet_balances::AccountData<u64>;
+    type OnNewAccount = ();
+    type OnKilledAccount = ();
+    type SystemWeightInfo = ();
+    type SS58Prefix = ();
+    type OnSetCode = ();
+    type MaxConsumers = ConstU32<16>;
+}
 
-    fn reverse(acc: u64) -> Result<MultiLocation, u64> {
-        Err(acc)
+parameter_types! {
+    pub const ExistentialDeposit: u64 = 10;
+}
+
+impl pallet_balances::Config for Test {
+    type MaxLocks = ();
+    type MaxReserves = ();
+    type ReserveIdentifier = [u8; 8];
+    type Balance = Balance;
+    type DustRemoval = ();
+    type RuntimeEvent = RuntimeEvent;
+    type ExistentialDeposit = ExistentialDeposit;
+    type AccountStore = System;
+    type WeightInfo = ();
+    type HoldIdentifier = ();
+    type FreezeIdentifier = ();
+    type MaxHolds = ConstU32<0>;
+    type MaxFreezes = ConstU32<0>;
+}
+
+impl pallet_assets::Config for Test {
+    type RuntimeEvent = RuntimeEvent;
+    type Balance = Balance;
+    type AssetId = AssetId;
+    type AssetIdParameter = AssetId;
+    type Currency = Balances;
+    type CreateOrigin = AsEnsureOriginWithArg<EnsureSigned<AccountId>>;
+    type ForceOrigin = EnsureRoot<AccountId>;
+    type AssetDeposit = ConstU64<10>;
+    type MetadataDepositBase = ConstU64<10>;
+    type MetadataDepositPerByte = ConstU64<1>;
+    type AssetAccountDeposit = ConstU64<10>;
+    type ApprovalDeposit = ConstU64<10>;
+    type StringLimit = ConstU32<50>;
+    type Freezer = ();
+    type Extra = ();
+    type RemoveItemsLimit = ConstU32<100>;
+    type CallbackHandle = ();
+    type WeightInfo = ();
+    #[cfg(feature = "runtime-benchmarks")]
+    type BenchmarkHelper = ();
+}
+
+pub struct MatchOnlyAsset<Asset>(PhantomData<Asset>);
+impl<Asset: Get<AssetId>> xcm_executor::traits::MatchesFungibles<AssetId, Balance>
+    for MatchOnlyAsset<Asset>
+{
+    fn matches_fungibles(
+        a: &MultiAsset,
+    ) -> core::result::Result<(AssetId, Balance), xcm_executor::traits::prelude::Error> {
+        use sp_runtime::traits::SaturatedConversion;
+        match a {
+            MultiAsset {
+                fun: Fungible(amount),
+                ..
+            } => Ok((Asset::get(), (*amount).saturated_into::<u64>())),
+            _ => Err(xcm_executor::traits::prelude::Error::AssetNotHandled),
+        }
     }
 }
 
 parameter_types! {
-    pub Ancestry: MultiLocation = Junction::Parachain(101).into();
-    pub UnitWeightCost: u64 = 10;
-    pub WeightPrice: (AssetId, u128, u128) = (Concrete(MultiLocation::parent()), 1_000_000, 1024);
+    pub const DummyCheckingAccount: AccountId = 0;
+
+    // AssetId used as a fungible for benchmarks
+    pub const TransactAssetId: u128 = 1;
+    pub const TransactAssetLocation: MultiLocation = MultiLocation { parents: 0, interior: X1(GeneralIndex(TransactAssetId::get())) };
 }
 
-pub struct AllAssetLocationsPass;
-impl ContainsPair<MultiAsset, MultiLocation> for AllAssetLocationsPass {
-    fn contains(_: &MultiAsset, _: &MultiLocation) -> bool {
-        true
+// Use ONLY assets as the asset transactor.
+pub type AssetTransactor = FungiblesAdapter<
+    Assets,
+    MatchOnlyAsset<TransactAssetId>,
+    AccountIdConverter,
+    AccountId,
+    NoChecking,
+    DummyCheckingAccount,
+>;
+
+parameter_types! {
+    /// Maximum number of instructions in a single XCM fragment. A sanity check against weight
+    /// calculations getting too crazy.
+    pub const MaxInstructions: u32 = 100;
+    pub const MaxAssetsIntoHolding: u32 = 64;
+
+    pub const WeightPrice: (xcm::latest::AssetId, u128, u128) = (Concrete(TransactAssetLocation::get()), 1_000_000, 1024);
+    pub const UnitWeightCost: u64 = 10;
+}
+
+pub struct XcmConfig;
+impl xcm_executor::Config for XcmConfig {
+    type RuntimeCall = RuntimeCall;
+    type XcmSender = DevNull;
+    type AssetTransactor = AssetTransactor;
+    type OriginConverter = ();
+    type IsReserve = ReserveAssetFilter;
+    type IsTeleporter = ();
+    type UniversalLocation = UniversalLocation;
+    type Barrier = AllowUnpaidExecutionFrom<Everything>;
+    type Weigher = xcm_builder::FixedWeightBounds<UnitWeightCost, RuntimeCall, MaxInstructions>;
+    type Trader = xcm_builder::FixedRateOfFungible<WeightPrice, ()>;
+    type ResponseHandler = DevNull;
+    type AssetTrap = ();
+    type AssetLocker = ();
+    type AssetExchanger = ();
+    type AssetClaims = ();
+    type SubscriptionService = ();
+    type PalletInstancesInfo = AllPalletsWithSystem;
+    type MaxAssetsIntoHolding = MaxAssetsIntoHolding;
+    type FeeManager = ();
+    type MessageExporter = ();
+    type UniversalAliases = Nothing;
+    type CallDispatcher = RuntimeCall;
+    type SafeCallFilter = Everything;
+}
+
+impl pallet_xcm_benchmarks::Config for Test {
+    type XcmConfig = XcmConfig;
+    type AccountIdConverter = AccountIdConverter;
+    fn valid_destination() -> Result<MultiLocation, BenchmarkError> {
+        let valid_destination: MultiLocation = X1(AccountId32 {
+            network: None,
+            id: [0u8; 32],
+        })
+        .into();
+
+        Ok(valid_destination)
+    }
+    fn worst_case_holding(_depositable_count: u32) -> MultiAssets {
+        let assets: Vec<MultiAsset> = vec![MultiAsset {
+            id: Concrete(MultiLocation::parent()),
+            fun: Fungible(u128::MAX),
+        }];
+        assets.into()
     }
 }
 
-#[cfg(feature = "runtime-benchmarks")]
-pub fn mock_worst_case_holding() -> MultiAssets {
-    let assets: Vec<MultiAsset> = vec![MultiAsset {
-        id: Concrete(MultiLocation::parent()),
-        fun: Fungible(u128::MAX),
-    }];
-    assets.into()
+impl pallet_xcm_benchmarks::generic::Config for Test {
+    type RuntimeCall = RuntimeCall;
+
+    fn worst_case_response() -> (u64, Response) {
+        let assets: MultiAssets = (Concrete(Here.into()), 100).into();
+        (0, Response::Assets(assets))
+    }
+
+    fn worst_case_asset_exchange() -> Result<(MultiAssets, MultiAssets), BenchmarkError> {
+        Err(BenchmarkError::Skip)
+    }
+
+    fn universal_alias() -> Result<(MultiLocation, Junction), BenchmarkError> {
+        Err(BenchmarkError::Skip)
+    }
+
+    fn export_message_origin_and_destination(
+    ) -> Result<(MultiLocation, NetworkId, Junctions), BenchmarkError> {
+        Err(BenchmarkError::Skip)
+    }
+
+    fn transact_origin_and_runtime_call() -> Result<(MultiLocation, RuntimeCall), BenchmarkError> {
+        Ok((
+            Default::default(),
+            frame_system::Call::remark_with_event { remark: vec![] }.into(),
+        ))
+    }
+
+    fn subscribe_origin() -> Result<MultiLocation, BenchmarkError> {
+        Ok(Default::default())
+    }
+
+    fn claimable_asset() -> Result<(MultiLocation, MultiLocation, MultiAssets), BenchmarkError> {
+        let assets: MultiAssets = (Concrete(Here.into()), 100).into();
+        let ticket = MultiLocation {
+            parents: 0,
+            interior: X1(GeneralIndex(0)),
+        };
+        Ok((Default::default(), ticket, assets))
+    }
+
+    fn unlockable_asset() -> Result<(MultiLocation, MultiLocation, MultiAsset), BenchmarkError> {
+        Err(BenchmarkError::Skip)
+    }
+}
+
+parameter_types! {
+    pub const CheckingAccount: Option<(u64, MintLocation)> = None;
+    pub const TrustedTeleporter: Option<(MultiLocation, MultiAsset)> = None;
+}
+
+impl pallet_xcm_benchmarks::fungible::Config for Test {
+    type TransactAsset = ItemOf<Assets, TransactAssetId, AccountId>;
+    type CheckedAccount = CheckingAccount;
+    type TrustedTeleporter = TrustedTeleporter;
+
+    fn get_multi_asset() -> MultiAsset {
+        let min_balance = 100u64;
+        let asset_location: MultiLocation = GeneralIndex(TransactAssetId::get()).into();
+
+        assert_ok!(Assets::force_create(
+            RuntimeOrigin::root(),
+            TransactAssetId::get(),
+            0u64,
+            true,
+            min_balance,
+        ));
+
+        MultiAsset {
+            id: Concrete(asset_location),
+            fun: Fungible((min_balance * 100).into()),
+        }
+    }
+}
+
+parameter_types! {
+    pub TrustedReserveLocation: MultiLocation = Parent.into();
+    pub TrustedReserveAsset: MultiAsset = MultiAsset { id: Concrete(TrustedReserveLocation::get()), fun: Fungible(1_000_000) };
+    pub TrustedReserve: Option<(MultiLocation, MultiAsset)> = Some((TrustedReserveLocation::get(), TrustedReserveAsset::get()));
+}
+
+impl fungible::Config for Test {
+    type TrustedReserve = TrustedReserve;
+}
+
+impl generic::Config for Test {}
+impl Config for Test {}
+
+pub fn new_test_ext() -> sp_io::TestExternalities {
+    let t = GenesisConfig {
+        ..Default::default()
+    }
+    .build_storage()
+    .unwrap();
+    t.into()
 }
